+++ src/main/java/org/openkuva/kuvabase/bwcj/data/entity/interfaces/transaction/ITransactionProposal.java
@@ -52,7 +52,7 @@ public interface ITransactionProposal {
 
     IOutput[] getOutputs();
 
-    int getAmount();
+    long getAmount();
 
     Object getMessage();
